third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/the-form-element/form-action-reflection-expected.txt

This is a testharness.js-based test.
PASS An action URL should be resolved relative to the document's base URL (= the document's URL in this case)
PASS An empty-string action content attribute should cause the IDL attribute to return the document's URL (= the document's base URL in this case)
FAIL A missing action content attribute should cause the IDL attribute to return the document's URL (= the document's base URL in this case)assert_equals: expected "http://web-platform.test:8001/html/semantics/forms/the-form-element/form-action-reflection.html" but got ""
Harness: the test ran to completion.
